+++ src/azure-cli/azure/cli/command_modules/vm/tests/latest/test_vm_commands.py
@@ -3304,6 +3304,17 @@ class VMGalleryImage(ScenarioTest):
         self.cmd('sig image-definition delete -g {rg} --gallery-name {gallery} --gallery-image-definition {image}')
         self.cmd('sig delete -g {rg} --gallery-name {gallery}')
 
+    @ResourceGroupPreparer(name_prefix='cli_test_gallery_specialized_', location='eastus2')
+    def test_gallery_specialized(self, resource_group):
+        self.kwargs.update({
+            'gallery': 'gallery1',
+            'image': 'image1'
+        })
+        self.cmd('sig create -g {rg} --gallery-name {gallery}', checks=self.check('name', '{gallery}'))
+        self.cmd('sig image-definition create -g {rg} --gallery-name {gallery} --gallery-image-definition {image} --os-type linux --os-state specialized -p publisher1 -f offer1 -s sku1',
+                 checks=[self.check('name', '{image}'), self.check('osState', 'Specialized')])
+
+
 # endregion